- CCManager: wx30 build fix
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@9741 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/ccmanager.cpp
@@ -63,13 +63,15 @@ namespace CCManagerHelper
         return !(dist < 0 || dist > stc->LinesOnScreen()); // caret is off screen
     }
 
+    // return a hash of a calltip context (to avoid storing strings of each calltip)
+    // used in m_CallTipChoiceDict and m_CallTipFuzzyChoiceDict
     static int CallTipToInt(const wxString& firstTip, int numPages)
     {
         int val = 33 * firstTip.Length() ^ numPages;
         for (wxString::const_iterator itr = firstTip.begin();
              itr != firstTip.end(); ++itr)
         {
-            val = 33 * val ^ (*itr);
+            val = 33 * val ^ static_cast<int>(*itr);
         }
         return val;
     }
